+++ changelog.md
@@ -1,3 +1,9 @@
+1.1.0 (2017-01-16)
+------------------
+
+- Add onNewEntityInstance event
+- Fix UnitOfWork::dispatchOnCreateEntityOverrideLocalValues() PHPDoc
+
 1.0.1 (2016-12-14)
 ------------------